chore(abstract-utxo): move implementation
- moves implementation from bitgo to abstract-utxo module

Ticket: BG-49417
