+++ modules/abstract-utxo/package.json
@@ -38,6 +38,17 @@
     ]
   },
   "dependencies": {
-    "@bitgo/sdk-core": "^1.0.1"
+    "@bitgo/blockapis": "^1.0.1",
+    "@bitgo/sdk-api": "^1.0.1",
+    "@bitgo/sdk-core": "^1.0.1",
+    "@bitgo/unspents": "^0.8.0",
+    "@bitgo/utxo-lib": "^2.2.2",
+    "@types/bluebird": "^3.5.25",
+    "@types/lodash": "^4.14.121",
+    "bip32": "^2.0.6",
+    "bitcoinjs-message": "^2.0.0",
+    "bluebird": "^3.5.3",
+    "lodash": "^4.17.14",
+    "superagent": "^3.8.3"
   }
 }
