+++ modules/abstract-utxo/src/config.ts
@@ -1,0 +1,8 @@
+// Supported cross-chain recovery routes. The coin to be recovered is the index, the valid coins for recipient wallets
+// are listed in the array.
+export const supportedCrossChainRecoveries = {
+  btc: ['bch', 'ltc', 'bsv'],
+  bch: ['btc', 'ltc', 'bsv'],
+  ltc: ['btc', 'bch', 'bsv'],
+  bsv: ['btc', 'ltc', 'bch'],
+};
