+++ modules/abstract-utxo/src/index.ts
@@ -1,0 +1,5 @@
+export * from './abstractUtxoCoin';
+export * from './config';
+export * from './recovery';
+export * from './replayProtection';
+export * from './sign';
