+++ modules/abstract-utxo/src/recovery/index.ts
@@ -1,0 +1,7 @@
+export * from './RecoveryProvider';
+export * from './backupKeyRecovery';
+export * from './baseApi';
+export * from './coingeckoApi';
+export * from './crossChainRecovery';
+export * from './mempoolApi';
+export * from './smartbitApi';
