+++ modules/abstract-utxo/src/sign.ts
