+++ modules/sdk-core/src/bitgo/utils/triple.ts
@@ -3,4 +3,3 @@ export type Triple<T> = [T, T, T];
 export function isTriple<T>(arr: T[]): arr is Triple<T> {
   return arr.length === 3;
 }
-
